Moved create_hom_snv_featuremap to featuremap_utils
